% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{hint}
\alias{hint}
\title{Show the hint for an exercise}
\usage{
hint(x)
}
\arguments{
\item{x}{An exercise.}
}
\value{
The exercise, invisibly.
}
\description{
Show the hint for an exercise
}
\examples{
x <- draw_exercise(topic = "nps", seed = 1)
hint(x)
}
\seealso{
\code{\link[=check_answer]{check_answer()}}, \code{\link[=reveal]{reveal()}}.

Other exercises:
\code{\link[=check_answer]{check_answer()}},
\code{\link[=reveal]{reveal()}}
}
\concept{exercises}
